+++ depthai-ros/CHANGELOG.rst
@@ -2,6 +2,11 @@
 Changelog for package depthai-ros
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+2.6.3 (2023-02-23)
+___________
+* Fix sensor name detection
+* Enable subpixel mode
+
 2.6.3 (2023-02-10)
 ___________
 * Camera calibration updates